backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
supabase==2.4.2
python-dotenv==1.0.1
pandas>=1.5.0
scikit-learn>=1.2.0
numpy>=1.23.0
